requirements.txt
scipy>=1.1.0
xarray>=0.10.7
matplotlib>=2.1.1
scikit-learn>=0.19.2
ipython>=5.0.0
numpydoc>=0.6.0
netCDF4>=1.3.1
dask>=0.16.0
toolz>=0.8.2
sns>=0.9.0